requirements.txt
scikit-learn
streamlit
pandas
numpy
plotly
seaborn
matplotlib
wordcloud
PyYAML
joblib
gdown
imbalanced-learn

